requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
rapidfuzz
easyocr
pillow
requests